--- __init.py__ ---
* Added support for Python 3.13 (which is the one the newest versions of Home Assistant are shipped with).

* Replaced the deprecated import from "homeassistant.core", which was causing a warning, with an import from "homeassistant.core_config".

-- ha_central_system.py ---

Replaced the deprecated import from "homeassistant.core", which was causing a warning, with an import from "homeassistant.core_config".

--- number.py ---

Replaced the deprecated import from "homeassistant.core", which was causing a warning, with an import from "homeassistant.core_config".

--- sensor.py ---

Replaced the deprecated import from "homeassistant.core", which was causing a warning, with an import from "homeassistant.core_config".